# backend/openapi/foloo-v1.yaml
openapi: 3.0.3
info:
  title: Foloo API
  version: 1.0.0-email-delivery
  description: Authenticated Foloo V1 API with offline sync, private media/PDF transfer, OAuth email connections and explicit follow-up delivery.
servers:
  - url: https://{apiId}.execute-api.us-east-1.amazonaws.com/v1
    variables:
      apiId:
        default: replace-after-deploy
security:
  - cognitoJwt: []
paths:
...
  /email/follow-ups:
    get:
      operationId: listEmailFollowUps
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
    post:
      operationId: prepareEmailFollowUp
      description: Creates an immutable preview from one owner-scoped Lead; it does not send.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PrepareFollowUpInput" }
      responses:
        "201": { $ref: "#/components/responses/DataResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
...
  /email/unsubscribe:
    security: []
    get:
      operationId: unsubscribeEmailRecipient
      description: Public, opaque and idempotent workspace-scoped opt-out.
      parameters:
        - {
            name: token,
            in: query,
            required: true,
            schema: { type: string, minLength: 32 },
          }
      responses:
        "200": { description: Opt-out recorded or already active. }
        "400": { description: Invalid token. }
  /email/templates:
    get:
      operationId: listEmailTemplates
      description: Lists the current seller's saved Event/Direct ES/EN templates.
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
    put:
      operationId: putEmailTemplate
      description: Saves one validated owner-scoped template; does not send email.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailTemplateInput" }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
  /workspace:
    get:
      operationId: getWorkspace
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
  /profile:
    get:
      operationId: getProfile
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
    put:
      operationId: putProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SellerProfileInput" }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
  /events:
    get:
      operationId: listEvents
      description: Full owner-scoped sync snapshot, including soft-deleted event tombstones (deletedAt). Clients hide tombstones from event pickers but retain them for associated Lead references.
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
    post:
      operationId: createEvent
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventInput" }
      responses:
        "201": { $ref: "#/components/responses/DataResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /events/{eventId}:
    parameters:
      - name: eventId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: updateEvent
      description: Optimistic, idempotent update of event name and dates.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventUpdateInput" }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteEvent
      description: Soft-deletes the event without deleting associated Leads or media.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventDeleteInput" }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /content:
    get:
      operationId: listContent
      description: Owner-scoped snapshot including soft-delete tombstones and temporary download URLs for available PDFs.
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
    post:
      operationId: createContent
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentInput" }
      responses:
        "201": { $ref: "#/components/responses/DataResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
  /content/{contentId}:
    parameters:
      - name: contentId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: updateContent
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentUpdateInput" }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteContent
      description: Creates a tombstone; does not delete the S3 object. Physical retention remains D-13.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentDeleteInput" }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /content/{contentId}/uploads:
    parameters:
      - name: contentId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: prepareContentUpload
      description: Returns a temporary S3 PUT authorization after backend size validation.
      responses:
        "201": { $ref: "#/components/responses/DataResponse" }
...
  /leads:
    get:
      operationId: listLeads
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
    post:
      operationId: createLead
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadInput" }
      responses:
        "201": { $ref: "#/components/responses/DataResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /leads/{leadId}:
    parameters:
      - name: leadId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: updateLead
      description: Updates only approved structured fields using optimistic revision control.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadUpdateInput" }
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
...
  /leads/{leadId}/media/uploads:
    parameters:
      - name: leadId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: prepareLeadMediaUpload
      description: Creates a short-lived authorization for one deterministic private S3 object.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeadMediaInput" }
      responses:
        "201": { $ref: "#/components/responses/DataResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
components:
  securitySchemes:
    cognitoJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito access token validated by API Gateway against the DEV issuer and App Client audience.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    DataResponse:
      description: Successful response
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: {}
    ErrorResponse:
      description: Stable safe error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    PrepareFollowUpInput:
      type: object
      additionalProperties: false
      required: [leadId, language]
      properties:
        id: { type: string, format: uuid }
        leadId: { type: string, format: uuid }
        language: { type: string, enum: [es, en] }
        subject:
          type: string
          maxLength: 200
          description: Frozen local subject; supplied together with both body fields.
        plainBody:
          type: string
          maxLength: 20000
          description: Frozen local plain body before the server-owned opt-out footer.
        htmlBody:
          type: string
          maxLength: 40000
          description: Frozen local HTML body before the server-owned opt-out footer.
...
    EmailTemplateInput:
      type: object
      additionalProperties: false
      required: [origin, language, subject, body, signature]
      properties:
        origin: { type: string, enum: [event, direct] }
        language: { type: string, enum: [es, en] }
        subject: { type: string, minLength: 1, maxLength: 500 }
        body: { type: string, minLength: 1, maxLength: 20000 }
        signature: { type: string, maxLength: 2000 }
    SellerProfileInput:
      type: object
      additionalProperties: false
      required: [name, company]
      properties:
        name: { type: string, minLength: 1, maxLength: 160 }
        company: { type: string, minLength: 1, maxLength: 160 }
        position: { type: string, nullable: true, maxLength: 160 }
        phone: { type: string, nullable: true, maxLength: 40 }
        photoUrl: { type: string, format: uri, nullable: true }
    EventInput:
      type: object
      additionalProperties: false
      required: [id, name, startsAt, endsAt]
      properties:
        id:
          {
            type: string,
            format: uuid,
            description: Stable UUID generated offline.,
          }
        name: { type: string, minLength: 1, maxLength: 160 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
    EventUpdateInput:
      type: object
      additionalProperties: false
      required: [revision, name, startsAt, endsAt]
      properties:
        revision: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 160 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
    EventDeleteInput:
      type: object
      additionalProperties: false
      required: [revision]
      properties:
        revision: { type: integer, minimum: 1 }
    LeadInput:
      type: object
      additionalProperties: false
      required: [id, capturedAt, origin, firstName, company, leadType, interest]
      properties:
        id:
          {
            type: string,
            format: uuid,
            description: Stable UUID generated offline.,
          }
        capturedAt: { type: string, format: date-time }
        origin: { type: string, enum: [event, direct] }
        eventId: { type: string, format: uuid, nullable: true }
        place: { type: string, nullable: true, maxLength: 240 }
        firstName: { type: string, minLength: 1, maxLength: 160 }
        lastName: { type: string, nullable: true, maxLength: 160 }
        position: { type: string, nullable: true, maxLength: 160 }
        company: { type: string, minLength: 1, maxLength: 160 }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true, maxLength: 40 }
        leadType: { type: string, enum: [customer, partner, supplier] }
        interest: { type: string, enum: [low, medium, high] }
        writtenNote: { type: string, nullable: true, maxLength: 10000 }
        commercialFolio: { type: string, nullable: true, maxLength: 80 }
        contentFileIds:
          description: Selected active Content IDs, frozen with display names by the server at capture time.
          type: array
          items: { type: string, format: uuid }
    LeadUpdateInput:
      type: object
      additionalProperties: false
      required: [revision, firstName, company, leadType, interest]
      properties:
        revision: { type: integer, minimum: 1 }
        firstName: { type: string, minLength: 1, maxLength: 160 }
        lastName: { type: string, nullable: true, maxLength: 160 }
        position: { type: string, nullable: true, maxLength: 160 }
        company: { type: string, minLength: 1, maxLength: 160 }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true, maxLength: 40 }
        leadType: { type: string, enum: [customer, partner, supplier] }
        interest: { type: string, enum: [low, medium, high] }
        writtenNote: { type: string, nullable: true, maxLength: 10000 }
        place:
          type: string
          nullable: true
          maxLength: 240
          description: Mutable only when the persisted Lead origin is direct.
    LeadMediaInput:
      type: object
      additionalProperties: false
      required: [id, kind, contentType, byteSize, capturedAt]
      properties:
        id: { type: string, format: uuid }
        kind:
          { type: string, enum: [business_card, reference_image, voice_note] }
        contentType: { type: string, enum: [image/jpeg, audio/m4a] }
        byteSize:
          type: integer
          minimum: 1
          maximum: 104857600
        capturedAt: { type: string, format: date-time }
        durationMs: { type: integer, minimum: 0, nullable: true }
        sha256: { type: string, pattern: "^[a-fA-F0-9]{64}$", nullable: true }
...
    ContentInput:
      type: object
      additionalProperties: false
      required: [id, displayName, fileName, byteSize, allEvents, eventIds]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string, minLength: 1, maxLength: 160 }
        fileName: { type: string, maxLength: 255, pattern: "\\.[Pp][Dd][Ff]$" }
        byteSize: { type: integer, minimum: 1, maximum: 25000000 }
        allEvents: { type: boolean }
        eventIds:
          type: array
          items: { type: string, format: uuid }
    ContentUpdateInput:
      type: object
      additionalProperties: false
      required: [revision, displayName, allEvents, eventIds]
      properties:
        revision: { type: integer, minimum: 1 }
        displayName: { type: string, minLength: 1, maxLength: 160 }
        allEvents: { type: boolean }
        eventIds:
          type: array
          items: { type: string, format: uuid }
    ContentDeleteInput:
      type: object
      additionalProperties: false
      required: [revision]
      properties:
        revision: { type: integer, minimum: 1 }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }